refactor: move BanUserButton component
